DCON-3558: Simplify AuditEvent migration to streaming copy+delete
Replace the multi-mode orchestrator (hourly partitions, ClickHouse
state table, PartitionWorker, MigrationVerifier, delete/reset flags)
with a single streaming cursor that copies each batch to ClickHouse
and then deletes it from MongoDB. The live collection itself is now
the progress ledger, so the audit_event_migration_state table and
its surrounding machinery are gone. Drop the now-unused helpers and
DDL, rehome the AuditEventTransformer tests alongside other
dataLayer transformer tests, and adjust the ClickHouse container
setup to reflect the removed DDL file.
